+++ service/src/main/java/mobi/nowtechnologies/server/service/PaymentPolicyService.java
@@ -55,7 +55,7 @@ public class PaymentPolicyService {
 		return paymentPolicyRepository.findOne(id.shortValue());
 	}
 
-	public PaymentPolicy getPaymentPolicy(final int operatorId, String paymentType, byte communityId){
+	public PaymentPolicy getPaymentPolicy(final int operatorId, String paymentType, int communityId){
 		Validate.notNull(paymentType, "The parameter paymentType is null");
         return paymentPolicyDao.getPaymentPolicy(operatorId, paymentType, communityId);
 	}
